+++ thermion_dart/native/include/material/grid_android.h
@@ -12,7 +12,7 @@ extern "C" {
 #endif
 
 #define GRID_GRID_OFFSET 0
-#define GRID_GRID_SIZE 41144
+#define GRID_GRID_SIZE 42135
 #define GRID_GRID_DATA (GRID_PACKAGE + GRID_GRID_OFFSET)
 
 #endif