+++ thermion_dart/native/include/material/image_vulkan.h
@@ -12,7 +12,7 @@ extern "C" {
 #endif
 
 #define IMAGE_IMAGE_OFFSET 0
-#define IMAGE_IMAGE_SIZE 30725
+#define IMAGE_IMAGE_SIZE 31223
 #define IMAGE_IMAGE_DATA (IMAGE_PACKAGE + IMAGE_IMAGE_OFFSET)
 
 #endif